[BUGFIX] Cast `$chars` argument of `fixed_lgd_cs()` to int
Since #97578, `GeneralUtility` is declared strict, potentially breaking
calls to `fixed_lgd_cs()` in at least the "List" module. The argument
`$chars` must be of type int, but sometimes it's a string which breaks
the call of `abs()`.

All calls to this method are adjusted, where required, and the value is
explicitly cast to int.

Resolves: #97887
Related: #97578
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75091
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>

+++ typo3/sysext/backend/Classes/Utility/BackendUtility.php
@@ -553,9 +553,9 @@ class BackendUtility
             if ($record['uid'] === 0) {
                 continue;
             }
-            $output = '/' . GeneralUtility::fixed_lgd_cs(strip_tags($record['title']), $titleLimit) . $output;
+            $output = '/' . GeneralUtility::fixed_lgd_cs(strip_tags($record['title']), (int)$titleLimit) . $output;
             if ($fullTitleLimit) {
-                $fullOutput = '/' . GeneralUtility::fixed_lgd_cs(strip_tags($record['title']), $fullTitleLimit) . $fullOutput;
+                $fullOutput = '/' . GeneralUtility::fixed_lgd_cs(strip_tags($record['title']), (int)$fullTitleLimit) . $fullOutput;
             }
         }
         if ($fullTitleLimit) {
@@ -1468,7 +1468,7 @@ class BackendUtility
     {
         // If $titleLength is not a valid positive integer, use BE_USER->uc['titleLen']:
         if (!$titleLength || !MathUtility::canBeInterpretedAsInteger($titleLength) || $titleLength < 0) {
-            $titleLength = static::getBackendUserAuthentication()->uc['titleLen'];
+            $titleLength = (int)static::getBackendUserAuthentication()->uc['titleLen'];
         }
         $titleOrig = htmlspecialchars($title);
         $title = htmlspecialchars(GeneralUtility::fixed_lgd_cs($title, $titleLength));
@@ -1742,7 +1742,7 @@ class BackendUtility
             $l = GeneralUtility::callUserFunction($_funcRef, $params, $null);
         }
         if ($fixed_lgd_chars && $l) {
-            return GeneralUtility::fixed_lgd_cs((string)$l, $fixed_lgd_chars);
+            return GeneralUtility::fixed_lgd_cs((string)$l, (int)$fixed_lgd_chars);
         }
         return $l;
     }
